pyproject.toml
[project]
name = "factorio-learning-environment"
version = "0.1.0"
description = "Factorio Learning Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "pydantic>=2.10.6",
    "tenacity>=9.0.0",
    "anthropic>=0.49.0",
    "openai>=1.66.3",
    "numpy>=2.2.3",
    "lupa>=2.4",
    "slpp>=1.2.3",
    "factorio-rcon-py>=2.1.3",
    "construct>=2.10.70",
    "pillow>=11.1.0",
    "scikit-image>=0.25.2",
    "psycopg2-binary>=2.9.10",
    "google-api-python-client>=2.164.0",
    "google-auth>=2.38.0",
    "rich>=14.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "env/src/models"]
